+++ Covenant/Models/Listeners/Listener.cs
@@ -4,7 +4,6 @@
 
 using System;
 using System.IO;
-using System.Linq;
 using System.Threading;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
@@ -13,11 +12,6 @@ using System.ComponentModel.DataAnnotations.Schema;
 using Newtonsoft.Json;
 
 using Covenant.Core;
-using Covenant.Models.Grunts;
-using NLog;
-using Covenant.API.Models;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore.Internal;
 
 namespace Covenant.Models.Listeners
 {
@@ -75,7 +69,7 @@ namespace Covenant.Models.Listeners
         public virtual CancellationTokenSource Start() { return null; }
         public virtual void Stop(CancellationTokenSource cancellationTokenSource) { }
 
-        protected string ListenerDirectory { get { return Common.CovenantListenersDirectory + this.GUID + Path.DirectorySeparatorChar; } }
+        protected string ListenerDirectory { get { return Common.CovenantListenersDirectory + Utilities.GetSanitizedFilename(this.GUID) + Path.DirectorySeparatorChar; } }
 
         // Listener|Action|ID|Name|GUID|BindAddress|BindPort|ConnectAddresses|ConnectPort|ProfileID|ListenerType|Status
         public string ToLog(LogAction action) => $"Listener|{action}|{this.Id}|{this.Name}|{this.GUID}|{this.BindAddress}|{this.BindPort}|{string.Join(",", this.ConnectAddresses)}|{this.ConnectPort}|{this.ProfileId}|{this.ListenerType.Name}|{this.Status}";